changed the filesnames to name of the comopent from the "main.jsx" convetorted all the ".jsx"  to ".js" files
