+++ src/shared/mainMenu/mainMenu.js
